requirements.txt
absl-py==1.4.0
astor==0.8.1
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
flatbuffers==1.12
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
termcolor==2.2.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
